boards/arm: Basic BSP structure for BLE Cell board
This commit adds the basic BSP structure for the nrf52840
based BLE Cell board. Most of the things in this commit are
okay and usable as-is, though the device tree file needs to
be modified per the board details.

Builds fine but RTT throws out a warning, which seems to be
coming from the 2.4 release. Will fix anyway.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>