modem: quectel-bg9x: Adding support for registering modem
Adding code which allows to register the Modem with the Zephyr
networking stack. At boot up, the Zephyr kernel will call up
"modem_init" routine to setup the modem, and then the
"modem_net_iface_init" to register the modem with the Zephyr
net stack.

Till now, we've added no code that talks to the BG9x modem or
powers it up etc. All code added till now is only the
infrastructure to support the modem driver.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>