modem: quectel-bg9x: Adding support for reading RSSI
Adding support for reading RSSI in the quectel driver. The
process of reading RSSI happens early on in the driver boot
process but takes place in a work so it doesn't block rest
of the boot process.

Command for getting the RSSI is the standard "AT+CSQ" which
returns the following,

  -> 0      means -113dBm or less
  -> 1      means -111dBm
  -> 2...30 means -109...-53dBm
  -> 31     means -51dBm or greater
  -> 99     means Not known or not detectable

We've added an algorithm to detect various states, and dry-run them
to ensure that correct RSSI is reported.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>