modem: quectel-bg9x: Adding support for reading IMEI & other static information.
This commit adds basic infrastructure to the driver to the driver to "setup modem".
This includes sending "AT" commands to the modem to ensure that its up and capable
of responding before doing anything else. Also, this commit adds support for reading
static information like IMEI, Model, Manufacturer etc. All of this information will
be presented to the user when they use the modem shell.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>