modem: quectel-bg9x: Adding code to power up the modem.
Till now, all code should work on BG95 & BG96 but this commit
won't as the power up sequence for both chips is different.
For example, the BG95 chip has RESET / PWRKEY pins internally
connected and requires the PWRKEY pin to be driven low for
500-1000 msec. Contrarily the BG96 chip doesn't have the RESET /
PWRKEY internally connected, and also requires different
sequence to power up.

This code follows the BG95 document. It drives,
-> PWRKEY high for 750msec which should power on the module.
-> PWRKEY low and then waits for a couple of seconds for UART
   to come up.

Eventually, we may need to add sequence for BG96 but not
something we need to worry about for now.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>