modem: quectel-bg9x: Getting RSSI
Add code to read RSSI at boot-up. Until we get valid RSSI, the
driver won't move forward. If we do, the driver will register
the RSSI work to keep checking if RSSI falls (in the background)
and proceed with doing other things.

The code in modem_setup is close, but it still needs some updates,
i.e. things like connecting with a given APN etc. But most of what's
required is already there.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>