modem: quectel-bg9x: Kick starting the implementation of Modem Socket APIs
This commit starts the implementation of Modem sockets API with the connect
function. Given an IP address, a port number and protocol type, the function
tells the modem to make connection with the given attributes.

The modem supports multiple sockets, but we have currently limited ourselves
to use only 1. This will be more than enough for testing / validating the driver,
and we will add support for multiple sockets later.

On the side, this commit also adds socket response handlers. These handlers are
invoked when the modem receives things like "OK" or "ERROR" from the modem.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>