modem: quectel-bg9x: Adding support for sendmsg and close
Filling up more and more modem socket dispatch table routines.
The ones that I've added now are sendmsg and close - The latter
is simply used to close an existing connection with a client,
while the former is an abstraction on top of sendto which
allows the higher level APIs to send multiple messages in a
single go.

By the looks of it, only the main APIs are left, i.e. Read and
Write - Most of everything else required to run a simple application
is done. Of course, there are other things like bind and listen which
we will eventually implement, but they are not critical for running
simple apps.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>