modem: quectel-bg9x: Support for data send
Actual support for sending data through the modem is added,
through all of this code still needs to be tested.

The workflow is relatively simple but async. Whenever a request
comes up to send data, we send "AT+QISEND" command to the modem.
If the modem responds to this command with a data prompt (>), we
write all data over UART and send a CTRL+Z. Then wait for the
modem to send data and respond to us with SEND OK or SEND FAIL.
If we fail to receive a data prompt, it means that the modem was
unable to send data to the client (because of some reason).

Except for reading data from the modem, most of the support seems
to be there but again, all needs to be tested.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>