dts: mtd: add support for a file system table in devicetree
Define a binding for zephyr,fstab which contains multiple child nodes
each of which identifies a file system type and configuration
parameters and associates it with a fixed partition.  The base entry
properties specify the mount point and other generic fs mount options,
while bindings specific to a file system type provide the
configuration parameters necessary to describe the file system.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>