drivers: sensor: Adding basic driver for TI BQ2589X
Adding basic driver for TI BQ2589X. This driver is not according to the
Zephyr Sensor API, but is still placed under sensor/ to decouple it from
the application. The user application can simply set CONFIG_BQ2589X=y,
include the bq2589x header file and make direct calls to the driver.

Note that the driver is specific to the "ABC board", in the sense that it
sets GPIO(s) which will only work on this board. We can probably work to
push all of this information out of the driver to the Device Tree file,
but I'm not sure about the value it will generate vs the time it will take.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>