drivers/sensor: Base driver for BQ2589x
Starting my attempt at a proper zephyr sensor driver for BQ2589x. This driver
is already working for us, but is not integrated properly with the zephyr
sensor subsystem. This commit kick-starts that attempt, though very little
real work is done yet.

For now, I've just added the base infrastructure necessary to build the
driver,
  west build -p auto -b contextualelectronics_abc samples/hello_world \
  -DCONFIG_I2C=y -DCONFIG_SENSOR=y -DCONFIG_BQ2589X=y

The driver builds but with warnings which will be fixed in the future
revisions.

One of the important things still to-do (and probably the most important
initially) is to define required device tree bindings for this battery
controller. I will probably go ahead and use whatever linux is using,
in addition to using the basic things like I2C etc.

Signed-off-by: Bilal Wasim <bilalwasim676@gmail.com>